requirements.txt
python-telegram-bot==21.5
aiohttp==3.9.5
redis[asyncio]>=5.0.0
apscheduler==3.10.4
python-dotenv==1.0.1
